+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Release 1.1.0
 Release 1.1.1
 
 - add `didChangeDependencies` callback to allow updating the value based on an `InheritedWidget`
+- add `updateShouldNotify` method to both `Provider` and `StatefulProvider`